+++ src/main/java/io/fastpix/sdk/SDKConfiguration.java
@@ -20,8 +20,8 @@ import io.fastpix.sdk.utils.Utils;
 public class SDKConfiguration {
 
     private static final String LANGUAGE = "java";
-    public static final String OPENAPI_DOC_VERSION = "1.0.0";
-    public static final String SDK_VERSION = "1.0.0";
+    public static final String OPENAPI_DOC_VERSION = "1.0.1";
+    public static final String SDK_VERSION = "1.0.1";
     public static final String GEN_VERSION = "2.788.7";
     private static final String BASE_PACKAGE = "io.fastpix.sdk";
     public static final String USER_AGENT =